+++ src/redux-ducks/chapters/index.js
@@ -16,7 +16,7 @@ import {
 } from "./actions";
 
 // ================================================================================
-// Reducers
+// Reducer
 // ================================================================================
 type State = $ReadOnly<{ [mangaId: number]: Array<ChapterType> }>;
