+++ src/redux-ducks/extensions/index.js
@@ -14,7 +14,7 @@ import {
 } from "./actions";
 
 // ================================================================================
-// Reducers
+// Reducer
 // ================================================================================
 type State = $ReadOnlyArray<ExtensionType>;
